chore(refactor): refactor taco package internals from character based to function based APIs
