+++ packages/taco/src/index.ts
@@ -1,9 +1,6 @@
 export { DkgPublicKey, ThresholdMessageKit } from '@nucypher/nucypher-core';
 
-export { Condition, Enrico, getPorterUri } from '@nucypher/shared';
-
-export { ThresholdDecrypter, ThresholdDecrypterJSON } from './cbd-recipient';
-
-export { CbdStrategy } from './cbd-strategy';
+// TODO: Re-xport condition types from @nucypher/shared
+export { Condition, getPorterUri } from '@nucypher/shared';
 
 export { taco } from './taco';
